Fixed model texture. Recompiled Assimp in Release (load time improved 2.8s -> 2s). Refactored Model. Copy dlls on build in CMake.

+++ project/include/renderer/master_renderer.hpp
@@ -37,7 +37,7 @@ public:
         global_light = light;
     }
 
-    void render(const Camera& camera, float time);
+    void render(const Camera& camera, double time);
 
 private:
     static glm::mat4 projection_matrix;
