+++ project/src/renderer/master_renderer.cpp
@@ -14,7 +14,7 @@ void MasterRenderer::calculateProjectionMatrix(float fov, float znear,
         glm::radians(fov), display_width / display_height, znear, zfar);
 }
 
-void MasterRenderer::render(const Camera& camera, float time) {
+void MasterRenderer::render(const Camera& camera, double time) {
     glEnable(GL_DEPTH_TEST);
     glClearColor(0.5f, 0.5f, 0.5f, 1.0f);
     glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
